MyActivity to test having multiple activities.
added static setActivity(Activity) method to ProcessThreadMessages so can update what the current activity is.  in the on create of all activities it should have ProcessThreadMessages.setActivity(this); near the beginning.

If we have time , we should change all of our views to activites to take advantage of build in android stuff